Finished moveGen function
Function generates a list of all pseudo legal moves in a position including castling and en Passant. Legality of moves will be checked by make move function immediately after making a move: If the white king is in check after making white's move it is an illegal move and needs to be unmade and the search terminated at that node.