Finished make move function
Finished make move function that updates entire Game State. Need to change something in the future about the 50 move draw rule, perhaps encode 0001 for general pawn pushes and not just double pawn pushes, as we need to check if there is a quiet move that isn't a pawn push for the 50 move draw rule. Also added a function to flip a bit in a bitboard using xor. Not totally necessary but it could be helpful down the line.